API debugger

+++ Raminagrobis.DAL/Prix_DAL.cs
@@ -10,13 +10,13 @@ namespace Raminagrobis.DAL
     {
         public int IDLignePanierG { get; private set; }
         public int IDFournisseur { get; private set; }
-        public int Prix { get; private set; }
+        public double Prix { get; private set; }
         public LignePanierG_DAL LignesG { get; private set; }
         public Fournisseur_DAL Fours { get; private set; }
 
-        public Prix_DAL(int idLignePanierG, int idFournisseur, int prix)
+        public Prix_DAL(int idLignePanierG, int idFournisseur, double prix)
             => (IDLignePanierG, IDFournisseur, Prix) = (idLignePanierG, idFournisseur, prix);
-        public Prix_DAL(LignePanierG_DAL lignesG, Fournisseur_DAL fours, int prix)
+        public Prix_DAL(LignePanierG_DAL lignesG, Fournisseur_DAL fours, double prix)
             => (LignesG, Fours, Prix) = (lignesG, fours, prix);
 
     }
